chore: refactor copy removing unused codes
